+++ LoxFramework/LoxFramework.csproj
@@ -53,6 +53,7 @@
     <Compile Include="AST\Expression.cs" />
     <Compile Include="AST\Statement.cs" />
     <Compile Include="Environment.cs" />
+    <Compile Include="Evaluating\AstPrinter.cs" />
     <Compile Include="LoxRunTimeException.cs" />
     <Compile Include="Extensions.cs" />
     <Compile Include="Interpreter.cs" />
