+++ UnitTests/UnitTests.csproj
@@ -46,7 +46,6 @@
     <Compile Include="..\SharedAssemblyInfo.cs">
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="LoxFramework\AstPrinter.cs" />
     <Compile Include="LoxFramework\InterpreterTests.cs" />
     <Compile Include="LoxFramework\TokenTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
